--- src/resources/js/features/StockControl.tsx
@@ -1,29 +1,0 @@
-import { useState } from 'react';
-import axios from 'axios';
-import { useInventoryItemState } from '@/hooks/InventoryItems';
-
-export const useBulkData = () => {
-// interface BulkAddResponse {
-//     success: boolean;
-//     message: string;
-// }
-
-    const {bulkData, setBulkData} = useInventoryItemState();
-    const [successMessage, setSuccessMessage] = useState(''); //消すかも
-
-    const DeliverhandleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-        event.preventDefault();
-        try {
-            const items = bulkData.split('\n').map(item => {
-                const [productName, modelNumber, location, inItem, remarks] = item.split(',').map(part => part.trim());
-                return { productName, modelNumber, location, inItem: parseInt(inItem), remarks };  //APIのリクエストに返す
-            });
-            const response = await axios.post('/api/items/bulk', { items });
-            setSuccessMessage(response.data.success);
-        } catch (error) {
-            console.error('Error posting bulk items:', error);
-        }
-    };
-
-    return { bulkData, setBulkData, DeliverhandleSubmit, successMessage};
-};
