+++ src/resources/js/functions/FetchItemsData.tsx
